Fix Enter-pressing to login and register
Not exactly sure about the internal cause of this, but somehow using
`v-on:keyup` was causing the call to `register` to be fired twice, which
would then call `toggleLoading` twice, therefore making it seem like the
submit button didn't get disabled at all (because it was disabled and
re-enabled really quickly). The solution to this is apparently just
using `v-on:keydown`.

Closes #21.