Fix incorrect login response
Calling POST /api/auth returns (if successful) the newly created session ID,
but somehow, when you logged in, it would normally respond with a `nil`
in place of the session ID, UNLESS you had previously set a session
cookie, in which case it would respond with a session ID... but the one
in your session cookie, which was outdated at that point. So what was
the reason for this?

Turns out we were using `Plug.Conn.put_session` to set the session
cookie with the session ID. That's all well and good, but the problem is
that, to formulate the JSON response, we pull the session ID out of that
session cookie with `Plug.Conn.get_session`! This would respond with
`nil` UNLESS you already have a session cookie submitted via the
response, in which case it would return that. Apparently the session
cookie itself is set as soon as the response is actually sent, and not
at the time you execute `Plug.Conn.put_session`.

The solution is simply to use a little bit of variable re-usage, and
directly use the session ID returned from the Ecto Repo call. It's a
simple solution for a very strange issue!